Boss_zrt
zrt怪，闪亮登场！走位会追击英雄，技能：天女散花

+++ LoneHero/Classes/Enemy.h
@@ -7,7 +7,7 @@ class Enemy :public cocos2d::Node
 {
 protected:
 	int blood;
-	float direction;
+	double direction;
 	int speed;
 	cocos2d::Sprite* enemy;
 	cocos2d::Sprite* hero;
